Allow building greybus subsystem protocols
In greybus networks, SVC is responsible for discovery of new nodes, and
disconnections. In some cases, (such as BeaglePlay cc1352), a Zephyr
firmware works as the SVC. The SVC is a greybus interface, but not a
greybus node. Thus it does not necessarily have a control port (and no
manifest).

The goal is to have most of SVC functions integrated into the module
here. However, for now, just allow building greybus module with greybus
node stuff disabled.

Signed-off-by: Ayush Singh <ayush@beagleboard.org>